Added some new mixer API features for Boomer (Sun's OSS implementation)
Added MIXF_DISABLED flag to mark mixer controls that are currently inactive
(should be shown grayed on the GUI). They can be read or changed but the
change will not have any effect until some mode change is done in the
hardware.

Also added MIXT_ENUM_MULTI that is like MIXT_ENUM but with support for
multiple active choices. The value field is a bit mask (31 available bits).
Note that this control type is reserved only for Boomer. OSS drivers should
not use it.