Added initial libossmix version
The libossmix library is intended to replace direct use of mixer
ioctl calls in ossxmix. The library makes it possible to access mixers
in remote systems in addition to the local ones. The ossmixd server program
included in the sources needs to be running in the remote system.

For the time being there is no password protection in the protocol. Also
the protocol works only between systems that have the same endianess. Finally
loading the mixer structure over a TCP connection takes too long time and
needs to be optimized.