Reduced memory footprint of GRC3 by fixing the quality setting to 3
At this moment The GRC3 quality setting (src_quality) is limited to 3 and
the lookup tables for the other quality levels are not compiled in. This
reduces the memory footprint of OSS by 220k.

The min/max quality settings are not configurable. However the configure
script can be modified to allow selection of GRC_MIN/MAX_QUALITY. The srcconf.c
program will then create local_config.h based on these environment variables.